Extract duplicate testing session in SPI tests
