+++ presto-spi/src/test/java/com/facebook/presto/spi/block/TestVariableWidthBlockEncoding.java
@@ -13,64 +13,16 @@
  */
 package com.facebook.presto.spi.block;
 
-import com.facebook.presto.spi.ConnectorSession;
-import com.facebook.presto.spi.PrestoException;
-import com.facebook.presto.spi.security.Identity;
-import com.facebook.presto.spi.type.TimeZoneKey;
 import com.facebook.presto.spi.type.Type;
 import io.airlift.slice.DynamicSliceOutput;
 import org.testng.annotations.Test;
 
-import java.util.Locale;
-import java.util.Optional;
-
-import static com.facebook.presto.spi.StandardErrorCode.INVALID_SESSION_PROPERTY;
-import static com.facebook.presto.spi.type.TimeZoneKey.UTC_KEY;
+import static com.facebook.presto.spi.block.TestingSession.SESSION;
 import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
-import static java.util.Locale.ENGLISH;
 import static org.testng.Assert.assertEquals;
 
 public class TestVariableWidthBlockEncoding
 {
-    private static final ConnectorSession SESSION = new ConnectorSession()
-    {
-        @Override
-        public String getQueryId()
-        {
-            return "test_query_id";
-        }
-
-        @Override
-        public Identity getIdentity()
-        {
-            return new Identity("user", Optional.empty());
-        }
-
-        @Override
-        public TimeZoneKey getTimeZoneKey()
-        {
-            return UTC_KEY;
-        }
-
-        @Override
-        public Locale getLocale()
-        {
-            return ENGLISH;
-        }
-
-        @Override
-        public long getStartTime()
-        {
-            return 0;
-        }
-
-        @Override
-        public <T> T getProperty(String name, Class<T> type)
-        {
-            throw new PrestoException(INVALID_SESSION_PROPERTY, "Unknown session property " + name);
-        }
-    };
-
     @Test
     public void testRoundTrip()
     {
